Add files via upload
The booking and revenue analysis dashboards provide a comprehensive overview of the hospitality business's performance. Out of a total room capacity of 233K, 135K bookings were made, with 75% confirmed and 25% cancelled, resulting in an occupancy rate of 58%. Mumbai leads in booking volume, while Elite rooms are the most booked category. Atliq Palace and Atliq Exotica are the top-performing properties in terms of bookings. On the revenue side, the business generated ₹1.71 billion in total revenue, of which ₹1.51 billion came from successful bookings, and ₹298.77 million was lost due to cancellations. Elite rooms and Atliq Exotica contributed the most to revenue, with Mumbai being the highest-earning city. The majority of revenue came through third-party platforms like MakeMyTrip and others, indicating reliance on external booking sources.